Update PC-TransGame.ini
Every game mode has it set to 20, cannot test to confirm if this is why the mode is broken. (HH works in private on live server, and is perfectly fine on test server)
Other guess is there is to much here and things that are not changed (OnlineGameSettings) need to be removed.
NumRequiredPlayers needs to stay due to conflicting Coalesced numbers from players.